Release v1
Version 1.
Simple POC script to exploit unprotected wordpress website. You give it a domain name and it will check if both REST API and xmlrpc.php are available to public. If so, the script will retrieve the user list and bruteforce each of them using the dictionnaries (none are provided in this repo).